68 acabado

+++ ProyectosIG_x64_VS2022/ProyectosIG_x64_VS2022/IG1App/Scene.h
@@ -51,7 +51,7 @@ protected:
 
 	CompoundEntity* inventedNode3_;
 
-	GLuint angle = 0;
+	float angle = 0;
 
 };
